% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpictorial-package.R
\docType{package}
\name{ggpictorial-package}
\alias{ggpictorial}
\alias{ggpictorial-package}
\title{ggpictorial: Makes pictorial charts with ggplot}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

Other contributors:
\itemize{
  \item Datasketch [copyright holder, funder]
}

}
\keyword{internal}
