% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs_icon.R
\name{vec_ptype2.hd_Icn}
\alias{vec_ptype2.hd_Icn}
\alias{vec_cast.hd_Icn}
\title{Define a method for combining two icon lists

#' @method vec_c hd_Icn
#' @export
vec_c.hd_Icn <- function(x, y, ...) {
message("combining")
Icn(c(vctrs::vec_data(x), vctrs::vec_data(y)))
#z <- c(vctrs::vec_data(x), vctrs::vec_data(y))
#obj_x <- attr(x, "object")
#obj_y <- attr(y, "object")
#new_Icn(z, object = c(obj_x, obj_y))
}}
\usage{
\method{vec_ptype2}{hd_Icn}(x, y, ...)

\method{vec_cast}{hd_Icn}(x, to, ...)
}
\description{
Define a method for combining two icon lists

#' @method vec_c hd_Icn
#' @export
vec_c.hd_Icn <- function(x, y, ...) {
message("combining")
Icn(c(vctrs::vec_data(x), vctrs::vec_data(y)))
#z <- c(vctrs::vec_data(x), vctrs::vec_data(y))
#obj_x <- attr(x, "object")
#obj_y <- attr(y, "object")
#new_Icn(z, object = c(obj_x, obj_y))
}
}
